Moved a bit of validation from server onto the client.
The validation of the phone number that was perfomed by the class
controllers.AuthenticationServlet.java when signing in is now performed
on the client through a submit handler on the form element specified
inside the newly added script signUp.js.
Also, when validation checks fail the alert div that used to be inserted
in the HTML file through a custom JSP tag is now added by code in the
signUp.js file.