Removed the code that added the menu options dinamically.
The menu options are added directly in the sidebarMenuInclude.jsp rather
than added to the session object inside AuthenticationServlet and then
added with a forEach jstl tag in the jsp.

This was first avoided because it creates repetition in the jsp but the
downside was that it was producing clutter in the AuthenticationServlet
class. Now we are focusing on refactoring the code so it's more
understandable and debuggable.