Mainly completly refactored AuthenticationServlet into 2 classes.
Renamed AuthenticationServlet into SignInVerificationServlet and refactored
the code that was left from AuthenticationServlet. Because of this I
changed the value of the action attribute of the form in signIn.jsp.

Added the original database script for the application becuse of the
intention to modify some aspects of it.

Small tweeks in signUp.jsp.