Fixed various issues.
Added sql statements for creating separate database and user.

Removed debugging output statements from some of the classes.

Modified persistence unit to work with a local database rather than the
remote used when building the app.

Modified application parameter identifying the root url of the app.

Downgraded the version of gson dependency to prevent error message when
starting the app.