Added support for testing the app using containers
Modified create.sql so that it allows connections to the database from
any host.

Added Dockerfiles for building database and application images.

Modified persistence unit by adding property to allow connections from the
container running the web application to the container running the
database. The previous property that allowed connections to the database
when building without containers was commented.

Added instructions to README detailing the process of building using
Docker.