First commit for the second version of this project
The purpose of this branch is to refactor and add new
functionality that wasn't implemented in the previous version of
the project due to time constraints.

As part of the refactoring process most of the files were
deleted and will be re added in later commits with refactoring
performed on them.

The schema of the database was changed adding more
constraints and as a result the entities of the application
change to reflect the database schema.