Added interface for encoding passwords
The main difference from the previous commit is the addition of the
PasswordEncoder interface that will be used for encoding and
checking passwords. For the time being the only implementation of
the interface is one that does not alter passwords but future
commits will contain implementations that can be used to securely
persist the passwords in the database.